requirements.txt
requests
Pillow
python-escpos
pyusb
psutil
qrcode[pil]
pilmoji
python-dotenv
